+++ sbi_stream/models/__init__.py
@@ -1,0 +1,22 @@
+"""Neural network models for Jeans GNN."""
+
+from .layers import GNN, GNNBlock, MLP, Transformer, MultiHeadAttentionBlock
+from .flows import build_flows
+from .lightning import GNNEmbedding, TransformerEmbedding, NPE, SequentialNPE
+from .utils import get_activation, configure_optimizers, build_embedding_loss
+
+__all__ = [
+    'GNN',
+    'GNNBlock',
+    'MLP',
+    'Transformer',
+    'MultiHeadAttentionBlock',
+    'GNNEmbedding',
+    'TransformerEmbedding',
+    'NPE',
+    'SequentialNPE',
+    'build_flows',
+    'get_activation',
+    'configure_optimizers',
+    'build_embedding_loss',
+]
