+++ sbi_stream/models/layers/__init__.py
@@ -2,12 +2,13 @@
 
 from .gnn import GNN, GNNBlock
 from .mlp import MLP
-from .transformer import Transformer, MultiHeadAttentionBlock
+from .transformer import TransformerA, TransformerB, MultiHeadAttentionBlock
 
 __all__ = [
     'GNN',
     'GNNBlock',
     'MLP',
-    'Transformer',
+    'TransformerA',
+    'TransformerB',
     'MultiHeadAttentionBlock',
 ]
