+++ sbi_stream/models/layers/gnn.py
