+++ sbi_stream/models/layers/transformer.py
@@ -1,7 +1,10 @@
-from typing import Optional
+
+from typing import Optional, Callable, Dict
 
 import torch
 import torch.nn as nn
+from torch.nn import TransformerEncoder, TransformerEncoderLayer
+from torch import Tensor
 
 
 class MultiHeadAttentionBlock(nn.Module):
@@ -80,8 +83,87 @@ class MultiHeadAttentionBlock(nn.Module):
         return x
 
 
-class Transformer(nn.Module):
-    """Decoder-only transformer for set modeling.
+class TransformerA(nn.Module):
+    """ Decoder-only Transformer. This is the original Transformer model used
+    in https://arxiv.org/abs/2512.07960v1.
+
+    Parameters
+    ----------
+    feat_input_size : int
+        The dimension of the input features (per token).
+    pos_input_size : int
+        The dimension of the input positional features.
+    feat_embed_size : int
+        The dimension of the projected feature embedding.
+    pos_embed_size : int
+        The dimension of the projected positional embedding.
+    nhead : int
+        The number of heads in the multihead attention modules.
+    num_encoder_layers : int
+        The number of sub-encoder-layers in the encoder.
+    sum_features : bool, optional
+        Whether to sum the features along the sequence dimension. Default: False
+    dim_feedforward : int
+        The dimension of the feedforward network model.
+    activation_fn : callable, optional
+        The activation function to use for the embedding layer. Default: None
+
+    """
+
+    def __init__(
+        self, feat_input_size: int, pos_input_size: int, feat_embed_size: int = 32,
+        pos_embed_size: int = 32, nhead: int = 4, num_encoder_layers: int = 4,
+        dim_feedforward: int = 128, sum_features: bool = False,
+        activation_name: str = None, activation_args: Optional[Dict] = None,
+    ) -> None:
+        super().__init__()
+        self.feat_embed_size = feat_embed_size
+        self.pos_embed_size = pos_embed_size
+        self.d_model = feat_embed_size + pos_embed_size
+        self.nhead = nhead
+        self.num_encoder_layers = num_encoder_layers
+        self.dim_feedforward = dim_feedforward
+        self.activation_fn = activation_fn
+
+        encoder_layer = TransformerEncoderLayer(
+            d_model=self.d_model, nhead=nhead, dim_feedforward=dim_feedforward,
+            batch_first=True)
+        self.transformer_encoder = TransformerEncoder(
+            encoder_layer, num_encoder_layers)
+        self.feat_embedding_layer = nn.Linear(feat_input_size, feat_embed_size)
+        self.pos_embedding_layer = nn.Linear(pos_input_size, pos_embed_size)
+
+    def forward(self, x: Tensor, pos: Tensor, padding_mask: Optional[Tensor] = None) -> Tensor:
+        """
+        x: (batch, seq, feat_input_size)
+        pos: (batch, seq, pos_input_size)
+        padding_mask: (batch, seq) boolean mask where True indicates padding
+        """
+        x = self.feat_embedding_layer(x)
+        pos = self.pos_embedding_layer(pos)
+        src = torch.cat((x, pos), dim=-1)
+        output = self.transformer_encoder(src, src_key_padding_mask=padding_mask)
+
+        # NOTE: dimension only works when batch_first=True
+        if padding_mask is None:
+            output = output.sum(dim=1)
+        else:
+            if not self.training:
+                # apply correct padding mask for evaluation
+                # this happens because with both self.eval() and torch.no_grad()
+                # the transformer encoder may shorten the output length to
+                # the max non-padded length in the batch
+                max_seq_len = output.shape[1]
+                padding_mask = padding_mask[:, :max_seq_len]
+            output = output.masked_fill(padding_mask.unsqueeze(-1), 0)
+            output = output.sum(dim=1)
+
+        return output
+
+
+class TransformerB(nn.Module):
+    """Decoder-only Transformer. This is an alternative Transformer model, copied
+    from https://github.com/trivnguyen/jgnn.
 
     Parameters
     ----------
