+++ sbi_stream/models/lightning/gnn_embedding.py
